Restart scheduler after running schedule-set command
